modified cifar project structur + updated gitignore
