+++ cifar10/models/structurer/MlpStructurer.py
